requirements.txt
requests==2.28.1
python-telegram-bot==13.15
environs==9.5.0
